chore: main.js파일 분할
